+++ fugue/src/test/java/io/atlassian/fugue/IterablesTest.java
@@ -114,6 +114,14 @@ public class IterablesTest {
     assertThat(rangeTo(1, 5, 2), is(contains(1, 3, 5)));
   }
 
+  @Test public void rangeToMaxValue() {
+    assertThat(rangeTo(Integer.MAX_VALUE - 1, Integer.MAX_VALUE), is(contains(Integer.MAX_VALUE - 1, Integer.MAX_VALUE)));
+  }
+
+  @Test public void rangeToMinValue() {
+    assertThat(rangeTo(Integer.MIN_VALUE + 1, Integer.MIN_VALUE), is(contains(Integer.MIN_VALUE + 1, Integer.MIN_VALUE)));
+  }
+
   @Test public void rangeUntilStep() {
     assertThat(rangeUntil(1, 5, 2), is(contains(1, 3)));
   }